Don't initialize auth on initial page load
We could synchronously initialize auth state for the initial render from Supabase, but this will produce hydration/render mismatch for the server.

+++ src/services/supabase/supabase-service.ts
@@ -80,10 +80,7 @@ export async function deleteUserSupabase(userId: string): Promise<void> {
       throw new Error('No Supabase service key.')
     } else {
       const supabaseServer = createSupabaseClient(supabaseServiceKey)
-      const { error } = await supabaseServer.auth.api.deleteUser(
-        userId,
-        supabaseServiceKey
-      )
+      const { error } = await supabaseServer.auth.api.deleteUser(userId)
       if (error) {
         throw new Error(`Error while deleting user: ${error.message}`)
       }
